pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapbase"
version = "1.0.0"
description = "SnapBase - SQL Bot for natural language to SQL query conversion"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["sql", "database", "nlp", "llm", "query-builder"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mysql-connector-python>=8.0.33,<9.0.0",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "tabulate>=0.9.0,<0.10.0",
    "cryptography>=41.0.0,<42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-cov>=4.0",
]
security = [
    "bandit>=1.7.5",
    "safety>=2.3.0",
]

[project.scripts]
snapbase = "main:main"

[tool.setuptools]
packages = ["app", "config", "db", "llm", "utils"]
py-modules = ["main"]

[project.urls]
Homepage = "https://github.com/yourusername/sql-bot-3"
Repository = "https://github.com/yourusername/sql-bot-3"
Issues = "https://github.com/yourusername/sql-bot-3/issues"
